requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
gunicorn>=21.2,<22.0
dj-database-url>=2.1,<2.2
python-dotenv>=1.0,<1.1
django-tailwind>=3.6,<3.7
whitenoise>=6.6,<6.7

# Testing
pytest-django>=4.7,<4.8
factory-boy>=3.3,<3.4
pytest-cov>=4.1,<4.2